ClusteringBenchmarks

Documentation for ClusteringBenchmarks.

ClusteringBenchmarks.ami

ClusteringBenchmarks.confusion_matrix

ClusteringBenchmarks.load_gagolewski

ClusteringBenchmarks.nca

ClusteringBenchmarks.ami — Method

score = ami(refclust, clust)
score = ami(C)

Compute the Adjusted Mutual Information (AMI) of a clustering clust with respect to the "ground truth" refclust. Alternatively supply the confusion matrix C (see confusion_matrix).

Note

AMI assumes hard-clustering.

Implements Eq. 24a in: Vinh, N.X., Epps, J. and Bailey, J., 2009. Information theoretic measures for clusterings comparison: is a correction for chance necessary?. In Proceedings of the 26th annual international conference on machine learning (pp. 1073-1080).

ClusteringBenchmarks.confusion_matrix — Method

C = confusion_matrix(refclust, clust)

Compute the confusion matrix of a clustering clust with respect to the "ground truth" refclust. Both refclust and clust are vectors of cluster labels (integers).

C[i, j] is the number of objects that are in cluster i in refclust and in cluster j in clust.

ClusteringBenchmarks.load_gagolewski — Method

data, labelsets = load_gagolewski(battery, dataset)

Load a dataset from the Gagolewski collection of data sets. data is a d × N matrix of Float64 values, where d is the dimensionality of the data set and N is the number of observations. labelsets is a vector of label-vectors (for each label-vector label = labelsets[i], label[j] is the cluster number assigned to the jth data point). Most data sets have a single label-vector, but some provide multiple label sets. Typically the first is the one provided by the original creator of the data set, and is the recommended choice for benchmarking.

Examples

using ClusteringBenchmarks
data, labelsets = load_gagolewski("wut", "x2")

ClusteringBenchmarks.nca — Method

score = nca(refclust, clust)
score = nca(C)

Compute the Normalised Clustering Accuracy (NCA) of a clustering clust with respect to the "ground truth" refclust. Alternatively supply the confusion matrix C (see confusion_matrix).

Note

The NCA assumes hard-clustering and requires that the number of clusters in refclust and clust be the same (C must be square). It scales poorly when there are many (>20) clusters.

Implements Eq 35 in: Gagolewski, M. (2023). Normalised clustering accuracy: An asymmetric external cluster validity measure (preprint). URL: https://arxiv.org/pdf/2209.02935.pdf, DOI: 10.48550/arXiv.2209.02935